refactor: reorganize server actions and improve schematic detail page
feat: add client-side components and caching for user data
fix: update import paths for server actions
perf: optimize schematic detail page with suspense and caching
docs: remove unused loading component
