--- app/[locale]/(main)/schematics/[id]/loading.tsx
@@ -1,5 +1,0 @@
-import DetailSkeleton from '@/components/skeleton/detail.skeleton';
-
-export default function Loading() {
-	return <DetailSkeleton />;
-}
