+++ app/[locale]/(main)/schematics/[id]/page.tsx
@@ -1,23 +1,20 @@
 import { Metadata } from 'next/dist/types';
-import React, { cache } from 'react';
+import React, { Suspense } from 'react';
 
-import ErrorScreen from '@/components/common/error-screen';
 import SchematicDetailCard from '@/components/schematic/schematic-detail-card';
+import DetailSkeleton from '@/components/skeleton/detail.skeleton';
 
-import { getCachedUser, serverApi } from '@/action/common';
+import { getCachedSchematic, getCachedUser } from '@/action/query';
 import env from '@/constant/env';
 import { Locale } from '@/i18n/config';
 import { getTranslation } from '@/i18n/server';
 import { getErrorMessage, isError } from '@/lib/error';
 import { generateAlternate } from '@/lib/i18n.utils';
 import { formatTitle } from '@/lib/utils';
-import { getSchematic } from '@/query/schematic';
 
 type Props = {
 	params: Promise<{ id: string; locale: Locale }>;
 };
-
-const getCachedSchematic = cache((id: string) => serverApi((axios) => getSchematic(axios, { id })));
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
 	const { id, locale } = await params;
@@ -55,12 +52,11 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
 }
 
 export default async function Page({ params }: Props) {
-	const { id } = await params;
-	const schematic = await getCachedSchematic(id);
+	const { id, locale } = await params;
 
-	if (isError(schematic)) {
-		return <ErrorScreen error={schematic} />;
-	}
-
-	return <SchematicDetailCard schematic={schematic} />;
+	return (
+		<Suspense fallback={<DetailSkeleton />}>
+			<SchematicDetailCard id={id} locale={locale} />
+		</Suspense>
+	);
 }
