+++ app/[locale]/(main)/servers/[id]/(dashboard)/pause-server-button.tsx
@@ -7,7 +7,7 @@ import Tran from '@/components/common/tran';
 import { Button } from '@/components/ui/button';
 import { toast } from '@/components/ui/sonner';
 
-import { revalidate } from '@/action/common';
+import { revalidate } from '@/action/server-action';
 import useClientApi from '@/hooks/use-client';
 import useQueriesData from '@/hooks/use-queries-data';
 import useServer from '@/hooks/use-server';
