+++ components/common/comment-section.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { Suspense, useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
