+++ components/schematic/schematic-detail-card.tsx
@@ -1,8 +1,5 @@
-'use client';
-
 import { Share2Icon } from 'lucide-react';
 import dynamic from 'next/dynamic';
-import { useParams } from 'next/navigation';
 
 import CopyButton from '@/components/button/copy.button';
 import DownloadButton from '@/components/button/download.button';
@@ -21,7 +18,7 @@ import {
 	DetailTitle,
 	Verifier,
 } from '@/components/common/detail';
-import DetailSwipeToNavigate from '@/components/common/detail-swipe-to-navigate';
+import ErrorScreen from '@/components/common/error-screen';
 import SizeCard from '@/components/common/size-card';
 import Tran from '@/components/common/tran';
 import LikeAndDislike from '@/components/like/like-and-dislike';
@@ -29,50 +26,45 @@ import ItemRequirementCard from '@/components/schematic/item-requirement-card';
 import { EllipsisButton } from '@/components/ui/ellipsis-button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 
+import { getSession } from '@/action/common';
+import { getCachedSchematic } from '@/action/query';
 import env from '@/constant/env';
-import { useSession } from '@/context/session.context';
-import useClientApi from '@/hooks/use-client';
-import useToastAction from '@/hooks/use-toast-action';
+import { Locale } from '@/i18n/config';
 import ProtectedElement from '@/layout/protected-element';
-import { getSchematicData, getSchematics } from '@/query/schematic';
-import { SchematicDetail } from '@/types/response/SchematicDetail';
-import { ItemPaginationQuery } from '@/types/schema/search-query';
+import { isError } from '@/lib/error';
 
 const DeleteSchematicButton = dynamic(() => import('@/components/schematic/delete-schematic.button'));
 const TakeDownSchematicButton = dynamic(() => import('@/components/schematic/take-down-schematic.button'));
 
 type SchematicDetailCardProps = {
-	schematic: SchematicDetail;
+	id: string;
+	locale: Locale;
 };
 
-export default function SchematicDetailCard({
-	schematic: {
-		id,
+export default async function SchematicDetailCard({ id, locale }: SchematicDetailCardProps) {
+	const schematic = await getCachedSchematic(id);
+	const session = await getSession();
+
+	if (isError(schematic)) {
+		return <ErrorScreen error={schematic} />;
+	}
+
+	const {
 		name,
 		description,
-		tags,
-		metadata: { requirements },
-		verifierId,
-		itemId,
+		downloadCount,
 		likes,
 		dislikes,
 		userId,
-		isVerified,
+		verifierId,
 		width,
 		height,
-		downloadCount,
 		createdAt,
-	},
-}: SchematicDetailCardProps) {
-	const axios = useClientApi();
-	const { session } = useSession();
-
-	const getData = useToastAction({
-		title: <Tran text="copying" />,
-		content: <Tran text="downloading-data" />,
-		action: async () => await getSchematicData(axios, id),
-	});
-	const { locale } = useParams();
+		tags,
+		itemId,
+		metadata: { requirements },
+		isVerified,
+	} = schematic;
 
 	const link = `${env.url.base}/${locale}/schematics/${id}`;
 	const copyContent = <Tran text="copied-name" args={{ name }} />;
@@ -83,69 +75,72 @@ export default function SchematicDetailCard({
 
 	return (
 		<Detail>
-			<DetailSwipeToNavigate paramSchema={ItemPaginationQuery} queryKey={['schematics']} queryFn={getSchematics}>
-				<CopyButton position="absolute" variant="ghost" data={link} content={link}>
-					<Share2Icon />
-				</CopyButton>
-				<DetailContent>
-					<DetailImage src={imageUrl} errorSrc={errorImageUrl} alt={name} />
-					<DetailHeader>
-						<DetailTitle className="border-b">{name}</DetailTitle>
-						<Tabs defaultValue="info" className="overflow-hidden flex flex-col">
-							<TabsList className="grid w-full grid-cols-2">
-								<TabsTrigger value="info">
-									<Tran text="info" />
-								</TabsTrigger>
-								<TabsTrigger value="comment">
-									<Tran text="comment" />
-								</TabsTrigger>
-							</TabsList>
-							<TabsContent value="info">
-								<DetailInfo>
-									<DetailAuthor authorId={userId} />
-									<Verifier verifierId={verifierId} />
-									<SizeCard size={{ width, height }} />
-									<DetailDescription>{description}</DetailDescription>
-									<CreatedAt createdAt={createdAt} />
-									<ItemRequirementCard requirements={requirements} />
-									<DetailTagsCard tags={tags} type="schematic" />
-								</DetailInfo>
-							</TabsContent>
-							<TabsContent value="comment">
-								<CommentSection itemId={itemId} />
-							</TabsContent>
-						</Tabs>
-						<DetailActions>
-							<CopyButton title={copyContent} data={getData} />
-							<DownloadButton href={downloadUrl} fileName={downloadName} count={downloadCount} />
-							<LikeAndDislike itemId={itemId} like={likes} dislike={dislikes} />
-							<EllipsisButton>
-								<ProtectedElement
-									session={session}
-									filter={{
-										all: [
-											{
-												any: [{ authorId: userId }, { authority: 'DELETE_SCHEMATIC' }],
-											},
-											isVerified,
-										],
-									}}
-								>
-									<TakeDownSchematicButton id={id} name={name} />
-								</ProtectedElement>
-								<ProtectedElement
-									session={session}
-									filter={{
-										any: [{ authorId: userId }, { authority: 'DELETE_SCHEMATIC' }],
-									}}
-								>
-									<DeleteSchematicButton variant="command" id={id} name={name} />
-								</ProtectedElement>
-							</EllipsisButton>
-						</DetailActions>
-					</DetailHeader>
-				</DetailContent>
-			</DetailSwipeToNavigate>
+			<CopyButton position="absolute" variant="ghost" data={link} content={link}>
+				<Share2Icon />
+			</CopyButton>
+			<DetailContent>
+				<DetailImage src={imageUrl} errorSrc={errorImageUrl} alt={name} />
+				<DetailHeader>
+					<DetailTitle className="border-b">{name}</DetailTitle>
+					<Tabs defaultValue="info" className="overflow-hidden flex flex-col">
+						<TabsList className="grid w-full grid-cols-2">
+							<TabsTrigger value="info">
+								<Tran text="info" />
+							</TabsTrigger>
+							<TabsTrigger value="comment">
+								<Tran text="comment" />
+							</TabsTrigger>
+						</TabsList>
+						<TabsContent value="info">
+							<DetailInfo>
+								<DetailAuthor authorId={userId} />
+								<Verifier verifierId={verifierId} />
+								<SizeCard size={{ width, height }} />
+								<DetailDescription>{description}</DetailDescription>
+								<CreatedAt createdAt={createdAt} />
+								<ItemRequirementCard requirements={requirements} />
+								<DetailTagsCard tags={tags} type="schematic" />
+							</DetailInfo>
+						</TabsContent>
+						<TabsContent value="comment">
+							<CommentSection itemId={itemId} />
+						</TabsContent>
+					</Tabs>
+					<DetailActions>
+						<CopyButton
+							title={copyContent}
+							data={{
+								url: `/schematics/${id}/data`,
+							}}
+						/>
+						<DownloadButton href={downloadUrl} fileName={downloadName} count={downloadCount} />
+						<LikeAndDislike itemId={itemId} like={likes} dislike={dislikes} />
+						<EllipsisButton>
+							<ProtectedElement
+								session={session}
+								filter={{
+									all: [
+										{
+											any: [{ authorId: userId }, { authority: 'DELETE_SCHEMATIC' }],
+										},
+										isVerified,
+									],
+								}}
+							>
+								<TakeDownSchematicButton id={id} name={name} />
+							</ProtectedElement>
+							<ProtectedElement
+								session={session}
+								filter={{
+									any: [{ authorId: userId }, { authority: 'DELETE_SCHEMATIC' }],
+								}}
+							>
+								<DeleteSchematicButton variant="command" id={id} name={name} />
+							</ProtectedElement>
+						</EllipsisButton>
+					</DetailActions>
+				</DetailHeader>
+			</DetailContent>
 		</Detail>
 	);
 }
