+++ components/server/kick.button.tsx
@@ -3,24 +3,24 @@
 import { BanIcon } from 'lucide-react';
 import { useState } from 'react';
 
-
-
 import Tran from '@/components/common/tran';
-import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogTitle, AlertDialogTrigger } from '@/components/ui/alert-dialog';
+import {
+	AlertDialog,
+	AlertDialogAction,
+	AlertDialogCancel,
+	AlertDialogContent,
+	AlertDialogTitle,
+	AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 
-
-
-import { revalidate } from '@/action/common';
+import { revalidate } from '@/action/server-action';
 import { useSocket } from '@/context/socket.context';
 import useMessage from '@/hooks/use-message';
 
-
-
 import { useQueryClient } from '@tanstack/react-query';
-
 
 type BanButtonProps = {
 	id: string;
