+++ components/tag/tag-container.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { Suspense, useMemo } from 'react';
 
 import TagCard from '@/components/tag/tag-card';
